CVSL-1725 Extracting out condition routes into their own controller
To make space for another licences endpoint
